\name{geneIDdata}
\alias{geneIDdata}
\docType{data}
\title{
Homologene data
}
\description{
Homologene data to support conversion of ENTREZ gene ID and gene name between human and mouse species.
}
\references{
http://www.ncbi.nlm.nih.gov/homologene
}
\examples{
data(geneIDdata)
}
\keyword{datasets}
