\name{GSCA-package}
\alias{GSCA-package}
\alias{GSCA}
\docType{package}
\title{Gene Set Co-expression Analysis}
\description{
  
}
\details{
\tabular{ll}{
Package: \tab GSCA\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-04-26\cr
License: \tab LGPL >= 2.0\cr
LazyLoad: \tab yes\cr
}

}
\author{
YounJeong Choi

Maintainer: YounJeong Choi <ychoi@biostat.wisc.edu>
}
\references{Choi and Kendziorski (2009)}
\keyword{GSCA}
