% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNW.R
\name{plotNW}
\alias{plotNW}
\title{The function to draw a network from a correlation matrix}
\usage{
plotNW(genes = NULL, cormatrix, node.de = NULL, ncolor = "red",
  ecolor = NULL, ewidth = 1, gtype = "neato", ...)
}
\arguments{
\item{genes}{Node names to be displayed in the network plot. If \code{NULL}
(default), \code{rownames(cormatrix)} is used.}

\item{cormatrix}{A symmetric correlation matrix to draw a network from.}

\item{node.de}{A vector of 0's and 1's. The "1" nodes are colored in
\code{ncolor}.}

\item{ncolor}{Node color for selected genes specified by \code{node.de}.}

\item{ecolor}{Edge color palette. The default is \code{bluered(201)}
implemented by \code{gplots}, which sets c("blue", "white", "red") for the
correlation of c(-1, 0, 1).}

\item{ewidth}{Edge width. Default is set to 1.}

\item{gtype}{Graph type (\code{layout}) to be passed to \code{plot.graph}.
One of \code{dot}, \code{neato}, \code{twopi}, \code{circo}, and \code{fdp}.
The default is \code{neato}.}

\item{\dots}{The rest of arguments are passed to \code{plot.graph}.}
}
\description{
This function is a custom wrapper of \code{plot.graph} implemented in
\code{Rgraphviz}. It draws a network from a given correlation matrix. Nodes
can be colored by DE information, and edges can be colored based on the
correlation magnitude and direction.
}
\examples{

data(LungCancer3)
GS <- LungCancer3$info$GSdef
GSdesc <- LungCancer3$info$Name

setid <- "GO:0019216"
gid <- GS[[setid]]
ss <- c("SERPINA3", "SOD1", "SCAP", "NPC2", "ADIPOQ", "PRKAA1", "AGT",
"PPARA", "BMP6", "BRCA1")

plotNW(genes = ss, cormatrix = cor(t(LungCancer3$data$Michigan[gid, 87:96]), use = "pairwise.complete.obs"), node.de = c(rep(1, 5), rep(0, 5)), ncolor = "yellow", ecolor = bluered(201), ewidth = 5, gtype = "circo")

}
\references{
Choi and Kendziorski (2009)
}
\seealso{
\code{\link{plot.graph}}, \code{\link{Rgraphviz}}
}
\author{
YounJeong Choi
}
\keyword{GSCA}
\keyword{network}
