% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSCA-package.R
\docType{package}
\name{GSCA-package}
\alias{GSCA-package}
\alias{GSCA}
\title{Gene Set Co-expression Analysis}
\description{
\describe{
\item{plotMNW}{Draws a multi-edge network from correlation matrices from multiple studies.}
\item{plotNW}{Draws a network from a correlation matrix.}
\item{singleDC}{gene set differential co-expression analysis}
\item{metaDI}{Calculates the dispersion index of the gene list}}
}
\references{
Choi and Kendziorski (2009)
}
\author{
YounJeong Choi

Maintainer: YounJeong Choi <ychoi@biostat.wisc.edu>
}
\keyword{GSCA}
